requirements.txt
langchain==0.3.21
langchain-community==0.0.27
langgraph==0.3.18
openai==1.68.2
python-dotenv==1.0.1
pyyaml==6.0.1 